+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Ability to pick no coauthors
 
 ## [v0.1.0](https://github.com/stefanlogue/meteor/releases/tag/v0.1.0) - 2023-11-05
 ### Added